--- components/Table/Table.module.css
@@ -1,115 +1,0 @@
-.container {
-  position: relative;
-  display: grid;
-  padding: 2.8rem;
-  background-color: var(--white);
-  border-radius: 0.8rem;
-}
-
-.title {
-  display: flex;
-  justify-content: space-between;
-}
-
-.title__text {
-  font-size: 2.4rem;
-  font-weight: 700;
-  color: var(--black1);
-}
-
-.pagecount {
-  display: flex;
-  gap: 2rem;
-  align-items: center;
-}
-
-.pagecount__text {
-  padding: 0.4rem;
-  width: 10rem;
-  font-size: 1.4rem;
-  color: var(--gray50);
-}
-
-.arrowbutton {
-  padding: 1rem;
-  border: 0.1rem solid var(--gray30);
-  border-radius: 1rem;
-}
-.arrowbutton:hover {
-  background-color: var(--gray30);
-}
-
-.invitebutton {
-  width: 10rem;
-  height: 3.2rem;
-  background-color: var(--violet);
-  color: var(--white);
-  font-size: 1.4rem;
-  font-weight: 500;
-  border-radius: 0.4rem;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 0.8rem;
-  grid-area: invite;
-}
-
-.tableindex {
-  display: grid;
-}
-
-.tableindex__item {
-  margin-top: 2.8rem;
-  margin-bottom: 0.8rem;
-  font-size: 1.6rem;
-  color: var(--gray50);
-}
-
-.row {
-  display: grid;
-  align-items: center;
-  height: 7rem;
-  font-size: 1.6rem;
-}
-.row:not(:first-of-type) {
-  border-top: 0.1rem solid var(--gray20);
-}
-
-.row__item {
-  color: var(--black1);
-  display: flex;
-  align-items: center;
-  gap: 1.2rem;
-}
-
-.button {
-  border: 0.1rem solid var(--gray30);
-  width: 8rem;
-  height: 3rem;
-  font-weight: 600;
-  color: var(--violet);
-  border-radius: 0.4rem;
-}
-
-.hidebutton {
-  position: absolute;
-  bottom: 0;
-  width: 100%;
-  display: flex;
-  gap: 1rem;
-  align-items: center;
-  justify-content: center;
-  height: 2rem;
-  font-weight: 600;
-}
-.hidebutton:hover {
-  background-color: var(--gray30);
-  border-bottom-right-radius: 0.4rem;
-  border-bottom-left-radius: 0.4rem;
-}
-
-.row__icon {
-  width: 0.8rem;
-  height: 0.8rem;
-  border-radius: 10rem;
-}
