+++ components/Table/TableList/TableList.module.css
@@ -1,0 +1,37 @@
+.button {
+  border: 0.1rem solid var(--gray30);
+  width: 8rem;
+  height: 3rem;
+  font-weight: 600;
+  color: var(--violet);
+  border-radius: 0.4rem;
+}
+
+.row {
+  display: grid;
+  align-items: center;
+  height: 7rem;
+  font-size: 1.6rem;
+}
+.row:not(:first-of-type) {
+  border-top: 0.1rem solid var(--gray20);
+}
+
+.row__item {
+  color: var(--black1);
+  display: flex;
+  align-items: center;
+  gap: 1.2rem;
+}
+
+.row__icon {
+  width: 0.8rem;
+  height: 0.8rem;
+  border-radius: 10rem;
+}
+
+@media screen and (max-width: 767px) {
+  .row {
+    font-size: 1.4rem;
+  }
+}
