+++ components/Table/TablePagenation/HideButton.module.css
@@ -1,0 +1,16 @@
+.hidebutton {
+  display: flex;
+  gap: 1rem;
+  align-items: center;
+  justify-content: center;
+  position: absolute;
+  bottom: 0;
+  width: 100%;
+  height: 2rem;
+  font-weight: 600;
+}
+.hidebutton:hover {
+  border-bottom-right-radius: 0.4rem;
+  border-bottom-left-radius: 0.4rem;
+  background-color: var(--gray30);
+}
